Move `getName` to `ASTUtil::getAttrName`
