+++ rewriter/utils/Utils.h
@@ -34,6 +34,11 @@ public:
 
     static ast::ExpressionPtr thunkBody(core::MutableContext ctx, ast::ExpressionPtr &node);
 
+    /// Fetch the name and loc of a symbol or string literal, that's suitable for an `attr`, `attr_reader`,
+    /// `attr_writer`, or `attr_accessor`
+    static std::pair<core::NameRef, core::LocOffsets> getAttrName(core::MutableContext ctx,
+                                                                  const ast::ExpressionPtr &name);
+
     ASTUtil() = delete;
 };
 } // namespace sorbet::rewriter
